feat(loader): modify loader decorator struct
